Fix for Jupiter Aggregator V6 trades
This fix removes duplicate SOL transfers which happen when an Associated Token Account is used, as seen for Jupiter Aggregator V6 trades.

We need to suppress all WSOL transfers which happen within the ATA, and leave only the initial SOL transfer to the account. This prevents the balance check from failing.