Added environment variable DEFITAXES_INSTANCE_PATH
This allows you to set the instance path which will be used by Flask, if not specified it will default to the existing "instance" folder.

Changed the flask application variable from app to application to be compatible with mod_wsgi if Apache is used.

Also added Mypy config and fixed errors. Adding static typing is still a work in progress.