Chain: Avalanche, BSC and Optimism moved to Routescan API
The Etherscan V2 API is giving the error "Free API access is temporarily unavailable due to unusually high network activity..." repeatedly for Optimism, BSC and Avalanche.

Moving these to use the Routescan V2 API instead.

Also re-enabled Boba as this chain is now supported on the Routescan API as well.